+++ acc_app_optimisation/envs/builtin_envs.py
@@ -13,3 +13,4 @@ import cern_awake_env.simulation
 import cern_leir_transfer_line_env
 import cern_sps_tune_env
 import cern_sps_zs_alignment_env
+import linac3_lebt_tuning